requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.7
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==3.2.2
mysql-connector-python==8.0.18
pip==9.0.1
pkg-resources==0.0.0
protobuf==3.10.0
pycodestyle==2.5.0
pytz==2019.3
requests==2.22.0
rope==0.14.0
setuptools==39.0.1
six==1.13.0
SQLAlchemy==1.3.11
urllib3==1.25.7
webargs==5.5.2
Werkzeug==0.16.0